upgrade deps, and consumedcapacity units are not longs!

+++ amazon/dynamodb/client/src/main/kotlin/org/http4k/connect/amazon/dynamodb/model/ConsumedCapacity.kt
@@ -5,10 +5,10 @@ import se.ansman.kotshi.JsonSerializable
 @JsonSerializable
 data class ConsumedCapacity(
     val TableName: TableName? = null,
-    val CapacityUnits: Long? = null,
+    val CapacityUnits: Double? = null,
     val GlobalSecondaryIndexes: Map<String, Capacity>? = null,
     val LocalSecondaryIndexes: Map<String, Capacity>? = null,
-    val ReadCapacityUnits: Long? = null,
+    val ReadCapacityUnits: Double? = null,
     val Table: Capacity? = null,
-    val WriteCapacityUnits: Long? = null
+    val WriteCapacityUnits: Double? = null
 )
